backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Coding Challenge Generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clerk-backend-api>=3.0.3",
    "fastapi>=0.115.13",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "svix>=1.68.0",
    "uvicorn>=0.34.3",
]
